% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz.R
\name{biv_bar_plot}
\alias{biv_bar_plot}
\title{A function to display or save density plots created on all categorical features.}
\usage{
biv_bar_plot(
  dataset,
  classVar,
  order = NULL,
  colors = NULL,
  barType = "dodge",
  loc = NULL
)
}
\arguments{
\item{dataset}{Name of the data frame object}

\item{classVar}{Name of the target feature}

\item{order}{A vector listing the target feature labels in the desired order. To use default order leave this parameter. Default value NULL.}

\item{colors}{A vector listing which color to use to represent which target feature label. To have the function pick color leave this parameter. Default value NULL.}

\item{barType}{A string value indicating bar plot type i.e. stack, dodge, fill.}

\item{loc}{A string with the directory where you want to save the plots. If no location is provided the plots will be created and displayed but not stored as image files.}
}
\value{
Returns three outputs:
\enumerate{
\item If loc = NULL, returns plots and displays in the RStudio Plots window,
\item If loc has location provided, creates and saves the plots. Doesn't display in RStudio.
}
}
\description{
Creates (or saves as png if asked) bar plots for all density features. Organize the bar plots based on target feature categories and desired barplot types of: dodge, stack, fill.
}
\examples{
library(dplyr)
library(ggplot2)
biv_bar_plot(dataset = iris \%>\%
mutate(sepal_width_cat = ifelse(Sepal.Width < mean(iris$Sepal.Width), 'Low', 'High')),
classVar = Species)

}
