% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check.packages}
\alias{check.packages}
\title{Load packages}
\usage{
check.packages(pkg, repo = "https://cran.rstudio.com/")
}
\arguments{
\item{pkg}{Vector of package names.}

\item{repo}{Repository mirror from where the package needs to be installed.}
}
\description{
Installs and loads multiple R packages.
}
\details{
\itemize{
\item Package names must be in String format
\item Repository default is cran.rstudio.com
}
}
\examples{

 library(tidyverse)

 check.packages(c('dplyr', 'ggplot2'))

}
