% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz.R
\name{qtile_plot}
\alias{qtile_plot}
\title{Ratio of target categories in binned numeric predictor}
\usage{
qtile_plot(data, x_feat, y_feat, quantile = 4)
}
\arguments{
\item{data}{Name of the data frame object}

\item{x_feat}{Name of the predictor of the interest}

\item{y_feat}{Name of the target feature}

\item{quantile}{Number of quantiles to be used to bin the numeric feature}
}
\value{
Returns three outputs:
\enumerate{
\item Quantile boundaries,
\item in each quantiles based bin,
\item A line plot showing ratio in different bins. The output of the function can be passed to plotly to generate interactive plot
}
}
\description{
Plot the ratio of categories of target labels inside different bins created based on quantiles for numeric features
}
\examples{

library(tidyverse)
qtile_plot(iris, Sepal.Length, Species, 4)


}
