% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine data set}
\format{
An object of class \code{data.frame} with 1599 rows and 12 columns.
}
\usage{
wine
}
\description{
Contains a simplified dataset to fit binary classifiers
}
\examples{
wine

}
\keyword{datasets}
